Replace Signup screen with new get started screen
