+++ src/components/auth/login/WelcomeScreen.js
@@ -35,6 +35,7 @@ import { PrivacyPolicy, PrivacyPolicyAndTerms, Support } from '../../webView/web
 import { createStackNavigator } from '../../appNavigation/stackNavigation'
 import ready from '../ready'
 import Auth from '../../auth/Auth'
+import GetStarted from '../GetStarted'
 import AuthTorus from '../../auth/torus/AuthTorus'
 
 // const log = logger.child({ from: 'Welcome' })
@@ -198,6 +199,7 @@ welcome.navigationOptions = {
 }
 
 const routes = {
+  GetStarted,
   Auth: AuthType,
   PrivacyPolicyAndTerms,
   PrivacyPolicy,
